Split enormous spawn.(h|c)pp into multiple files
Aside from hygiene factors, at least the linker on our FreeBSD build
node could not handle the large object file generated from spawn.cpp.
