+++ libvast/src/system/spawn_archive.cpp
@@ -11,30 +11,34 @@
  * contained in the LICENSE file.                                             *
  ******************************************************************************/
 
-#pragma once
+#include "vast/system/spawn_archive.hpp"
 
-#include <iostream>
-#include <memory>
-#include <string>
+#include <caf/actor.hpp>
+#include <caf/actor_cast.hpp>
+#include <caf/config_value.hpp>
+#include <caf/expected.hpp>
+#include <caf/local_actor.hpp>
 
-#include "caf/fwd.hpp"
+#include "vast/defaults.hpp"
+#include "vast/filesystem.hpp"
+#include "vast/si_literals.hpp"
+#include "vast/system/archive.hpp"
+#include "vast/system/spawn_arguments.hpp"
 
-#include "vast/expected.hpp"
+using namespace vast::binary_byte_literals;
 
-namespace vast::detail {
+namespace vast::system {
 
-expected<std::unique_ptr<std::ostream>>
-make_output_stream(const std::string& output, bool is_uds = false);
+maybe_actor spawn_archive(caf::local_actor* self, spawn_arguments& args) {
+  namespace sd = vast::defaults::system;
+  if (!args.empty())
+    return unexpected_arguments(args);
+  auto segments = get_or(args.options, "global.segments", sd::segments);
+  auto mss = 1_MiB
+             * get_or(args.options, "global.max-segment-size",
+                      sd::max_segment_size);
+  auto a = self->spawn(archive, args.dir / args.label, segments, mss);
+  return caf::actor_cast<caf::actor>(a);
+}
 
-expected<std::unique_ptr<std::ostream>>
-make_output_stream(const caf::config_value_map& options);
-
-expected<std::unique_ptr<std::istream>>
-make_input_stream(const std::string& input, bool is_uds = false);
-
-expected<std::unique_ptr<std::istream>>
-make_input_stream(const caf::config_value_map& options);
-
-} // namespace vast::detail
-
-
+} // namespace vast::system
