+++ libvast/src/system/spawn_importer.cpp
@@ -11,50 +11,27 @@
  * contained in the LICENSE file.                                             *
  ******************************************************************************/
 
-#pragma once
+#include "vast/system/spawn_importer.hpp"
 
-#include <memory>
+#include <caf/actor.hpp>
+#include <caf/expected.hpp>
 
-#include <caf/fwd.hpp>
+#include "vast/defaults.hpp"
+#include "vast/detail/unbox_var.hpp"
+#include "vast/system/importer.hpp"
+#include "vast/system/node.hpp"
+#include "vast/system/spawn_arguments.hpp"
 
 namespace vast::system {
 
-// -- classes ------------------------------------------------------------------
-
-class application;
-class configuration;
-class default_application;
-class export_command;
-class import_command;
-class indexer_manager;
-class indexer_stage_driver;
-class node_command;
-class partition;
-class pcap_reader_command;
-class pcap_writer_command;
-class remote_command;
-class sink_command;
-class source_command;
-class start_command;
-
-// -- structs ------------------------------------------------------------------
-
-struct node_state;
-struct query_statistics;
-struct spawn_arguments;
-
-// -- templates ----------------------------------------------------------------
-
-template <class Reader>
-class reader_command;
-
-template <class Writer>
-class writer_command;
-
-// -- aliases ------------------------------------------------------------------
-
-using node_actor = caf::stateful_actor<node_state>;
-using partition_ptr = caf::intrusive_ptr<partition>;
-
+maybe_actor spawn_importer(node_actor* self, spawn_arguments& args) {
+  if (!args.empty())
+    return unexpected_arguments(args);
+  // FIXME: Notify exporters with a continuous query.
+  return self->spawn(importer, args.dir / args.label,
+                     caf::get_or(self->system().config(),
+                                 "vast.table-slice-size",
+                                 defaults::system::table_slice_size));
+}
 
 } // namespace vast::system
