+++ libvast/src/system/spawn_profiler.cpp
@@ -11,50 +11,35 @@
  * contained in the LICENSE file.                                             *
  ******************************************************************************/
 
-#pragma once
+#include "vast/system/spawn_profiler.hpp"
 
-#include <memory>
+#include <caf/actor.hpp>
+#include <caf/expected.hpp>
 
-#include <caf/fwd.hpp>
+#include "vast/config.hpp"
+#include "vast/detail/unbox_var.hpp"
+#include "vast/error.hpp"
+#include "vast/system/node.hpp"
+#include "vast/system/spawn_arguments.hpp"
 
 namespace vast::system {
 
-// -- classes ------------------------------------------------------------------
-
-class application;
-class configuration;
-class default_application;
-class export_command;
-class import_command;
-class indexer_manager;
-class indexer_stage_driver;
-class node_command;
-class partition;
-class pcap_reader_command;
-class pcap_writer_command;
-class remote_command;
-class sink_command;
-class source_command;
-class start_command;
-
-// -- structs ------------------------------------------------------------------
-
-struct node_state;
-struct query_statistics;
-struct spawn_arguments;
-
-// -- templates ----------------------------------------------------------------
-
-template <class Reader>
-class reader_command;
-
-template <class Writer>
-class writer_command;
-
-// -- aliases ------------------------------------------------------------------
-
-using node_actor = caf::stateful_actor<node_state>;
-using partition_ptr = caf::intrusive_ptr<partition>;
-
+maybe_actor spawn_profiler([[maybe_unused]] caf::local_actor* self,
+                           [[maybe_unused]] spawn_arguments& args) {
+#ifdef VAST_HAVE_GPERFTOOLS
+  return make_error(ec::unspecified, "not compiled with gperftools");
+#else // VAST_HAVE_GPERFTOOLS
+  if (!args.empty())
+    return unexpected_arguments(args);
+  auto resolution = args.opt("global.resolution", size_t{1});
+  auto secs = std::chrono::seconds(resolution);
+  auto prof = self->spawn(profiler, args.dir / args.label, secs);
+  if (args.opt("global.cpu", false))
+    anon_send(prof, start_atom::value, cpu_atom::value);
+  if (args.opt("global.heap", false))
+    anon_send(prof, start_atom::value, heap_atom::value);
+  return prof;
+#endif // VAST_HAVE_GPERFTOOLS
+}
 
 } // namespace vast::system
